backend (1)/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.2
Flask-CORS==4.0.0
bcrypt==4.0.1
PyTTSX3==2.90
google-cloud-speech==2.21.0
gunicorn==21.2.0
python-dotenv==1.0.0
marshmallow==3.20.1
flask-marshmallow==0.15.0
PyMySQL==1.1.0
Flask-Migrate==4.0.5
SQLAlchemy-JSONField
marshmallow-sqlalchemy